% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dateToTimeList}
\alias{dateToTimeList}
\title{Convert a time object to a list}
\usage{
dateToTimeList(value)
}
\arguments{
\item{value}{A time object}
}
\value{
A list with the hour, minute and second components
}
\description{
Convert a time object to a list
}
\keyword{internal}
