% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropNulls}
\alias{dropNulls}
\title{Drop NULL values from vector/lists}
\usage{
dropNulls(x)
}
\arguments{
\item{x}{A vector or list}
}
\value{
A vector or list with all the NULL items removed
}
\description{
Drop NULL values from vector/lists
}
