% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Round a time object to the nearest minute
From ?is.integer}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A time object}

\item{tol}{The tolerance for rounding}
}
\value{
A time object rounded to the nearest minute
}
\description{
Round a time object to the nearest minute
From ?is.integer
}
\keyword{internal}
