% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roundTime}
\alias{roundTime}
\title{Round a time object to the nearest minute}
\usage{
roundTime(time, minutes)
}
\arguments{
\item{time}{A time object}

\item{minutes}{The number of minutes to round to}
}
\value{
A time object rounded to the nearest minute
}
\description{
Round a time object to the nearest minute
}
