% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shinyInputLabel}
\alias{shinyInputLabel}
\title{Create a label tag for a given input}
\usage{
shinyInputLabel(inputId, label = NULL)
}
\arguments{
\item{inputId}{The `input` slot}

\item{label}{The label text}
}
\value{
A label tag
}
\description{
Create a label tag for a given input
}
