% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-time.R
\docType{package}
\name{shinyTime-package}
\alias{shinyTime}
\alias{shinyTime-package}
\title{shinyTime: A Time Input Widget for Shiny}
\description{
Provides a time input widget for Shiny. This widget allows intuitive time input in the '[hh]:[mm]:[ss]' or '[hh]:[mm]' (24H) format by using a separate numeric input for each time component. The interface with R uses date-time objects. See the project page for more information and examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://burgerga.github.io/shinyTime/}
  \item \url{https://github.com/burgerga/shinyTime}
  \item Report bugs at \url{https://github.com/burgerga/shinyTime/issues}
}

}
\author{
\strong{Maintainer}: Gerhard Burger \email{burger.ga@gmail.com} (\href{https://orcid.org/0000-0003-1062-5576}{ORCID})

}
\keyword{internal}
