% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-time.R
\docType{package}
\name{shinyTime}
\alias{shinyTime}
\alias{shinyTime-package}
\title{shinyTime: A timeInput widget for Shiny}
\description{
Provides a time input widget for inputting time in the %H:%M:%S format.
}

