% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-time.R
\name{shinyTimeExample}
\alias{shinyTimeExample}
\title{Run a simple example app using the shinyTime functionality}
\usage{
shinyTimeExample()
}
\description{
This functions runs a very simple app demonstrating the shinyTime functionality.
}
\seealso{
Other shinyTime.functions: \code{\link{timeInput}},
  \code{\link{updateTimeInput}}
}

