% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timeListToDate}
\alias{timeListToDate}
\title{Convert a list to a time object}
\usage{
timeListToDate(value)
}
\arguments{
\item{value}{A list with the hour, minute and second components}
}
\value{
A time object
}
\description{
Convert a list to a time object
}
\keyword{internal}
