% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timeStringToDate}
\alias{timeStringToDate}
\title{Convert a string to a time object}
\usage{
timeStringToDate(string)
}
\arguments{
\item{string}{A string with the time in the format "HH:MM:SS"}
}
\value{
A time object
}
\description{
Convert a string to a time object
}
