% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shinyInputLabel}
\alias{shinyInputLabel}
\title{Create a label tag for a given input}
\usage{
shinyInputLabel(inputId, label = NULL, control = FALSE)
}
\arguments{
\item{inputId}{The input ID}

\item{label}{The label text}

\item{control}{Whether the label is for a control (e.g., a checkbox)}
}
\value{
A label tag
}
\description{
Create a label tag for a given input
}
