% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-time.R
\docType{package}
\name{shinyTime}
\alias{shinyTime}
\alias{shinyTime-package}
\title{shinyTime: A Time Input Widget for Shiny}
\description{
Provides a time input widget for Shiny. This widget allows intuitive time input in the
\code{[hh]:[mm]:[ss]} (24H) format by using a separate numeric input for each part
of the time. The interface with R uses \code{\link{DateTimeClasses}} objects.
}

