% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-time.R
\name{shinyTimeExample}
\alias{shinyTimeExample}
\title{Show the shinyTime example app}
\usage{
shinyTimeExample()
}
\description{
Run a simple shiny app demonstrating the shinyTime functionality.
}
\seealso{
Other shinyTime functions: 
\code{\link{timeInput}()},
\code{\link{updateTimeInput}()}
}
\concept{shinyTime functions}
