% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-time.R
\name{timeInput}
\alias{timeInput}
\title{Create a time input}
\usage{
timeInput(inputId, label, value = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{The desired time value. Must be a instance of \code{\link{DateTimeClasses}}.}
}
\description{
Creates a time widget that consists of three separate numeric inputs for respectively the hours,
minutes, and seconds. The input and output values of the time widget are instances of
\code{\link{DateTimeClasses}}, these can be converted to and from character strings with
\code{\link{strptime}} and \code{\link{strftime}}. For a simple example app see
\code{\link{shinyTimeExample}}.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
  # Default value is 00:00:00
  timeInput("time1", "Time:"),

  # Set to current time
  timeInput("time2", "Time:", value = Sys.time()),

  # Set to custom time
  timeInput("time3", "Time:", value = strptime("12:34:56", "\%T"))
)

shinyApp(ui, server = function(input, output) { })
}

}
\seealso{
\code{\link{strptime}}, \code{\link{strftime}}

Other shinyTime.functions: \code{\link{shinyTimeExample}},
  \code{\link{updateTimeInput}}
}

